requirements.txt
cohere
python-dotenv 